requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
BASC-py4chan==0.6.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
discord==1.0.1
discord.py==1.7.1
disputils==0.1.3
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
jwcrypto==0.8
multidict==5.1.0
oauth2client==4.1.3
Pillow==8.2.0
praw==7.2.0
prawcore==2.0.0
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-decouple==3.4
python-jwt==3.3.0
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.58.0
yarl==1.6.3
